ajustes reglas de negocio, estado venta

+++ src/main/java/com/project/Project/project/model/ReversionVentaDTO.java
@@ -8,7 +8,7 @@ public class ReversionVentaDTO {
 
     @NotNull(message = "El ID de venta no puede ser nulo.")
     @Min(value = 1, message = "El ID de venta debe ser un número positivo.")
-    private Long idVenta;
+    private int idVenta;
 
     @NotBlank(message = "El motivo de reversión no puede estar vacío.")
     private String motivoReversion;
@@ -19,17 +19,17 @@ public class ReversionVentaDTO {
     public ReversionVentaDTO() {
     }
 
-    public ReversionVentaDTO(Long idVenta, String motivoReversion, ArrayList devuelto) {
+    public ReversionVentaDTO(int idVenta, String motivoReversion, ArrayList devuelto) {
         this.idVenta = idVenta;
         this.motivoReversion = motivoReversion;
         this.devuelto = devuelto;
     }
 
-    public Long getIdVenta() {
+    public int getIdVenta() {
         return idVenta;
     }
 
-    public void setIdVenta(Long idVenta) {
+    public void setIdVenta(int idVenta) {
         this.idVenta = idVenta;
     }
